Revive bugzilla.rhbugzilla.RHBugzilla import path
I've had a few reports about apps breaking without it.

We can't easily move RHBugzilla to the existing rhbugzilla.py due
to cyclic dependency issues.

Rename rhbugzilla.py to _rhconverters.py, and make rhbugzilla.py just
import RHBugzilla from oldclasses.py.

Add tests to ensure we don't regress in the future
